+++ run_clang_format.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-clang-format -style=file -i src/wsjcpp_sql_builder.h
-clang-format -style=file -i src/wsjcpp_sql_builder.cpp