+++ src/main.cpp
@@ -1,20 +1,48 @@
+/**********************************************************************************
+ * MIT License
+ *
+ * Copyright (c) 2025-2026 Evgenii Sopov <mrseakg@gmail.com>
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ *all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ *
+ * Official Source Code: https://github.com/sea5kg/sea5kg-sql-builder
+ *
+ ***********************************************************************************/
+
 #include <iostream>
-#include <wsjcpp_sql_builder.h>
+#include <sea5kg_sql_builder.h>
 
-int main(int argc, const char* argv[]) {
+int main(int argc, const char *argv[]) {
+  sea5kg::sql_builder builder;
+  // clang-format off
+  builder
+    .selectFrom("users")
+      .column("id")
+      .column("name")
+      .column("pass")
+    .where()
+    .equal("pass", "64d878e22cd5046d569bf826869034caf5feab24")
+  ;
+  // clang-format on
 
-    wsjcpp::SqlBuilder builder;
-        builder.selectFrom("users")
-        .column("id")
-        .column("name")
-        .column("pass")
-        .where()
-            .equal("pass", "64d878e22cd5046d569bf826869034caf5feab24")
-    ;
+  std::string sql = builder.sql();
+  std::cout << sql << std::endl;
 
-    std::string sql = builder.sql();
-    std::cout << sql << std::endl;
-
-    return 0;
+  return 0;
 }
-